move all DB logic to seapate package; making more generic
